[skip-list] Some spring cleaning, fixing a bug in `erase`.

+++ RandomHeight.h
@@ -6,5 +6,5 @@ class RandomHeight {
 
     private:
         int maxLevel;
-        int probability;
+        float probability;
 };
